doen

+++ src/utilities/localDb.ts
@@ -1,7 +1,6 @@
 import { IStatus, IUrl } from "../Interface";
 
 // key of localStorage
-
 const KEY = "urls";
 
 const parseData = (data: string | null): IUrl[] => {
@@ -18,6 +17,13 @@ const parseData = (data: string | null): IUrl[] => {
 
 const stringifyData = (data: IUrl | IUrl[]): string => {
   return JSON.stringify(data);
+};
+
+const isLinkAlreadyExist = (link: string) => {
+  const allLink = getAllData();
+  return allLink.find(
+    (single) => single.link === link || single.shortedLink === link
+  );
 };
 
 // get all data from db
@@ -61,10 +67,8 @@ const addSingleData = (newData: IUrl): IStatus<IUrl[]> => {
   };
 
   // check is url already exits
-  const isUrlAlreadyExist = previousData.find(
-    (single) => single.link === newData.link
-  );
-  if (!isUrlAlreadyExist) {
+  const checkIsLinkAlreadyExist = isLinkAlreadyExist(newData.link);
+  if (!checkIsLinkAlreadyExist) {
     const allData = [newData, ...previousData];
     localStorage.setItem(KEY, stringifyData(allData));
     addStatus.data = allData;
@@ -81,44 +85,40 @@ const updateSingleData = (link: string, updatableLink: string) => {
   const previousData = getAllData();
 
   // default status is not found
-  let addStatus: IStatus<IUrl[]> = {
+  let updateStatus: IStatus<IUrl[]> = {
     status: 400,
     message: "Link not found!",
     data: previousData,
   };
+  // check is url already exits
+  const checkIsLinkAlreadyExist = isLinkAlreadyExist(updatableLink);
 
-  let newUpdatedAllData: IUrl[] = [];
-  for (const single of previousData) {
-    // checking does this link already exist
-    if (single.link === updatableLink) {
-      addStatus.message = "Url Already exist";
-      addStatus.status = 400;
-      newUpdatedAllData = previousData;
-      break;
-    }
-
-    if (single.link === link) {
-      addStatus.message = "Link successfully updated";
-      addStatus.status = 200;
-      newUpdatedAllData = [
-        ...newUpdatedAllData,
-        { link: updatableLink, shortedLink: single.shortedLink },
-      ];
-    } else {
-      newUpdatedAllData = [...newUpdatedAllData, single];
-    }
+  if (checkIsLinkAlreadyExist) {
+    updateStatus.message = "Link Already Exist!";
+    return updateStatus;
   }
+
+  const newUpdatedAllData = previousData.map((single) => {
+    if (single.link === link) {
+      updateStatus.message = "Link successfully updated!";
+      updateStatus.status = 200;
+      return { link: updatableLink, shortedLink: single.shortedLink };
+    } else {
+      return single;
+    }
+  });
   localStorage.setItem(KEY, stringifyData(newUpdatedAllData));
 
   // update status data
-  addStatus.data = newUpdatedAllData;
-  return addStatus;
+  updateStatus.data = newUpdatedAllData;
+
+  return updateStatus;
 };
 
 const deleteDataByLink = (link: string) => {
   const previousData = getAllData();
 
-  let addStatus: IStatus<IUrl[]> = {
+  let deleteStatus: IStatus<IUrl[]> = {
     status: 400,
     message: "Link not found!",
     data: previousData,
@@ -128,16 +128,18 @@ const deleteDataByLink = (link: string) => {
     if (single.link !== link) {
       return true;
     }
-    addStatus.status = 200;
-    addStatus.message = "Link successfully deleted!";
+    deleteStatus.status = 200;
+    deleteStatus.message = "Link successfully deleted!";
     return false;
   });
-  addStatus.data = filterData;
-  if (addStatus.status === 200) {
+
+  deleteStatus.data = filterData;
+
+  if (deleteStatus.status === 200) {
     localStorage.setItem(KEY, stringifyData(filterData));
   }
 
-  return addStatus;
+  return deleteStatus;
 };
 
 const localDb = {
